requirements.txt
aiogram==3.17.0
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
alembic==1.14.1
pydantic-settings==2.7.1

